get radius to int

+++ SMAPExam/app/src/main/java/com/grp16/itsmap/smapexam/service/LocationParam.java
@@ -23,7 +23,7 @@ public class LocationParam {
         return longitude;
     }
 
-    public double getRadius() {
+    public int getRadius() {
         return radius;
     }
